2.1pre19
{
    "filename": "links-2.1pre19.tar.gz",
    "md5": "5dba73edeaf3357829e270f64944adcc",
    "sha1": "09a0dfb83fdc3c23f2b64a74a17fb281bbb0935f",
    "sha256": "c51f4b5b28f56293f437a79305055be7f8adaa7b645a3494a0a015c0249fed94"
}

+++ bfu.c
@@ -94,7 +94,7 @@ void freeml(struct memory_list *ml)
 }
 
 #ifndef G
-#define txtlen strlen
+#define txtlen (int)strlen
 #else
 #define txtlen(x) (F ? g_text_width(bfu_style_wb, x) : (int)strlen(x))
 #endif
